disque nu = 0 jaja

+++ database/migrations/2024_09_24_192736_create_comprobantes_table.php
@@ -35,7 +35,7 @@ return new class extends Migration {
             $table->string('resultado_asientos')->nullable();
             $table->integer('sin_afectacion')->nullable();
             $table->integer('cuenta_contrapartida')->nullable();
-            $table->integer('numerounico')->nullable();
+            $table->BigInteger('numerounico')->nullable();
             $table->timestamps();
         });
     }
